allowing observation to be of any type

+++ src/policies/greedy.jl
@@ -46,7 +46,7 @@ end
 # loop over all actions.
 # The one with smallest expected entropy is best
 # We know this depends on sensor
-function action(m::SearchDomain, x::Vehicle, o::Float64, f::DF, p::GreedyPolicy)
+function action(m::SearchDomain, x::Vehicle, o, f::DF, p::GreedyPolicy)
     best_mi = -Inf
     best_a = (0.0, 0.0, 0.0)
     for a in p.actions
